refactor: s3 upload 모듈 세분화
S3Client : 실제 이미지를 업로드하는 모듈
S3Service 이미지 업로드 관련 모든 기능을 담당
